services/mcp-travel-knowledge/pyproject.toml
[project]
name = "mcp-travel-knowledge"
version = "0.1.0"
description = "MCP server: travel knowledge retrieval (Weaviate-backed)"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.6",
  "weaviate-client>=4.0",
  "httpx>=0.27",
  "requests>=2.28",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "jsonschema>=4.20",
  "httpx>=0.27",
  "ruff>=0.6",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
asyncio_mode = "auto"
